requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
absl-py==0.13.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.7.4.post0
    # via fsspec
alembic==1.7.3
    # via optuna
appnope==0.1.2
    # via ipython
astunparse==1.6.3
    # via tensorflow
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   cmd2
    #   jsonschema
autopage==0.4.0
    # via cliff
backcall==0.2.0
    # via ipython
bleach==4.1.0
    # via nbconvert
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via aiohttp
charset-normalizer==2.0.6
    # via requests
clang==5.0
    # via tensorflow
click==8.0.1
    # via
    #   nltk
    #   norby
    #   sacremoses
    #   typer
cliff==3.9.0
    # via optuna
cmaes==0.8.2
    # via optuna
cmd2==2.2.0
    # via cliff
colorama==0.4.4
    # via cmd2
colorlog==6.4.1
    # via optuna
configparser==5.0.2
    # via MMVAE-Hub (setup.py)
cycler==0.10.0
    # via matplotlib
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via
    #   jupyter-client
    #   nbconvert
filelock==3.0.12
    # via
    #   huggingface-hub
    #   transformers
flatbuffers==1.12
    # via tensorflow
git+https://github.com/VLL-HD/FrEIA.git
    # via MMVAE-Hub (setup.py)
fsspec[http]==2021.8.1
    # via pytorch-lightning
future==0.18.2
    # via pytorch-lightning
gast==0.4.0
    # via tensorflow
google-auth==1.35.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
greenlet==1.1.1
    # via sqlalchemy
grpcio==1.40.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.1.0
    # via tensorflow
huggingface-hub==0.0.17
    # via transformers
idna==3.2
    # via
    #   requests
    #   yarl
imageio==2.9.0
    # via MMVAE-Hub (setup.py)
ipython==7.27.0
    # via MMVAE-Hub (setup.py)
ipython-genutils==0.2.0
    # via nbformat
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via nbconvert
joblib==1.0.1
    # via
    #   nltk
    #   sacremoses
    #   scikit-learn
jsonschema==3.2.0
    # via nbformat
jupyter-client==7.0.3
    # via nbclient
jupyter-core==4.8.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.1.2
    # via nbconvert
keras==2.6.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.3.2
    # via matplotlib
mako==1.1.5
    # via alembic
markdown==3.3.4
    # via tensorboard
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
matplotlib==3.4.3
    # via MMVAE-Hub (setup.py)
matplotlib-inline==0.1.3
    # via ipython
mistune==0.8.4
    # via nbconvert
git+https://github.com/Jimmy2027/MODUN.git
    # via MMVAE-Hub (setup.py)
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
nbclient==0.5.4
    # via nbconvert
nbconvert==6.1.0
    # via MMVAE-Hub (setup.py)
nbformat==5.1.3
    # via
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
nltk==3.6.2
    # via MMVAE-Hub (setup.py)
git+https://github.com/Jimmy2027/NORBY.git
    # via MMVAE-Hub (setup.py)
numpy==1.19.5
    # via
    #   cmaes
    #   freia
    #   h5py
    #   imageio
    #   keras-preprocessing
    #   matplotlib
    #   norby
    #   opt-einsum
    #   optuna
    #   pandas
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
    #   torchmetrics
    #   torchvision
    #   transformers
oauthlib==3.1.1
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
optuna==2.9.1
    # via MMVAE-Hub (setup.py)
packaging==21.0
    # via
    #   bleach
    #   huggingface-hub
    #   optuna
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
pandas==1.3.3
    # via MMVAE-Hub (setup.py)
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pbr==5.6.0
    # via
    #   cliff
    #   stevedore
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.2
    # via
    #   MMVAE-Hub (setup.py)
    #   imageio
    #   matplotlib
    #   torchvision
prettytable==2.2.0
    # via cliff
prompt-toolkit==3.0.20
    # via ipython
protobuf==3.18.0
    # via
    #   tensorboard
    #   tensorflow
ptyprocess==0.7.0
    # via pexpect
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydeprecate==0.3.1
    # via pytorch-lightning
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pymongo==3.12.0
    # via
    #   MMVAE-Hub (setup.py)
    #   modun
pyparsing==2.4.7
    # via
    #   cliff
    #   matplotlib
    #   packaging
pyperclip==1.8.2
    # via cmd2
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytorch-lightning==1.4.7
    # via MMVAE-Hub (setup.py)
pytz==2021.1
    # via pandas
pyyaml==5.4.1
    # via
    #   cliff
    #   optuna
    #   pytorch-lightning
    #   transformers
pyzmq==22.3.0
    # via jupyter-client
regex==2021.8.28
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.26.0
    # via
    #   fsspec
    #   huggingface-hub
    #   norby
    #   requests-oauthlib
    #   tensorboard
    #   transformers
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.7.2
    # via google-auth
sacremoses==0.0.45
    # via transformers
scikit-learn==0.24.2
    # via sklearn
scipy==1.7.1
    # via
    #   freia
    #   optuna
    #   scikit-learn
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   bleach
    #   cycler
    #   google-auth
    #   google-pasta
    #   grpcio
    #   jsonschema
    #   keras-preprocessing
    #   python-dateutil
    #   sacremoses
    #   tensorflow
sklearn==0.0
    # via MMVAE-Hub (setup.py)
sqlalchemy==1.4.23
    # via
    #   alembic
    #   optuna
stevedore==3.4.0
    # via cliff
subprocess-tee==0.3.4
    # via norby
tensorboard==2.6.0
    # via
    #   MMVAE-Hub (setup.py)
    #   pytorch-lightning
    #   tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorflow==2.6.0
    # via MMVAE-Hub (setup.py)
tensorflow-estimator==2.6.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
testpath==0.5.0
    # via nbconvert
threadpoolctl==2.2.0
    # via scikit-learn
tokenizers==0.10.3
    # via transformers
torch==1.9.0
    # via
    #   MMVAE-Hub (setup.py)
    #   freia
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==0.5.1
    # via pytorch-lightning
torchvision==0.10.0
    # via MMVAE-Hub (setup.py)
tornado==6.1
    # via jupyter-client
tqdm==4.62.2
    # via
    #   MMVAE-Hub (setup.py)
    #   huggingface-hub
    #   nltk
    #   optuna
    #   pytorch-lightning
    #   sacremoses
    #   transformers
traitlets==5.1.0
    # via
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
transformers==4.10.2
    # via MMVAE-Hub (setup.py)
typer==0.4.0
    # via MMVAE-Hub (setup.py)
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   huggingface-hub
    #   pytorch-lightning
    #   tensorflow
    #   torch
urllib3==1.26.6
    # via requests
wcwidth==0.2.5
    # via
    #   cmd2
    #   prettytable
    #   prompt-toolkit
webencodings==0.5.1
    # via bleach
werkzeug==2.0.1
    # via tensorboard
wget==3.2
    # via
    #   MMVAE-Hub (setup.py)
    #   modun
wheel==0.37.0
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
